// frontend/src/api/generated/souls-directory/souls-directory.ts
export type searchApiV1SoulsDirectorySearchGetResponse200 = {
  data: SoulsDirectorySearchResponse;
  status: 200;
};

export type searchApiV1SoulsDirectorySearchGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type searchApiV1SoulsDirectorySearchGetResponseSuccess =
  searchApiV1SoulsDirectorySearchGetResponse200 & {
    headers: Headers;
  };
export type searchApiV1SoulsDirectorySearchGetResponseError =
  searchApiV1SoulsDirectorySearchGetResponse422 & {
    headers: Headers;
  };

export type searchApiV1SoulsDirectorySearchGetResponse =
  | searchApiV1SoulsDirectorySearchGetResponseSuccess
  | searchApiV1SoulsDirectorySearchGetResponseError;
// ...
export type getMarkdownApiV1SoulsDirectoryHandleSlugGetResponse200 = {
  data: SoulsDirectoryMarkdownResponse;
  status: 200;
};

export type getMarkdownApiV1SoulsDirectoryHandleSlugGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type getMarkdownApiV1SoulsDirectoryHandleSlugGetResponseSuccess =
  getMarkdownApiV1SoulsDirectoryHandleSlugGetResponse200 & {
    headers: Headers;
  };
export type getMarkdownApiV1SoulsDirectoryHandleSlugGetResponseError =
  getMarkdownApiV1SoulsDirectoryHandleSlugGetResponse422 & {
    headers: Headers;
  };

export type getMarkdownApiV1SoulsDirectoryHandleSlugGetResponse =
  | getMarkdownApiV1SoulsDirectoryHandleSlugGetResponseSuccess
  | getMarkdownApiV1SoulsDirectoryHandleSlugGetResponseError;